unet/requirements.txt
albumentations==1.3.1
matplotlib==3.7.1
numpy==1.24.2
Pillow==10.2.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
